Add files via upload
Abstract: The growth of the global population coupled with a decline in natural resources, farmland,
and the increase in unpredictable environmental conditions leads to food security is becoming a major
concern for all nations worldwide. These problems are motivators that are driving the agricultural
industry to transition to smart agriculture with the application of the Internet of Things (IoT) and
big data solutions to improve operational efficiency and productivity. The IoT integrates a series
of existing state-of-the-art solutions and technologies, such as wireless sensor networks, cognitive
radio ad hoc networks, cloud computing, big data, and end-user applications. This study presents
a survey of IoT solutions and demonstrates how IoT can be integrated into the smart agriculture
sector. To achieve this objective, we discuss the vision of IoT-enabled smart agriculture ecosystems
by evaluating their architecture (IoT devices, communication technologies, big data storage, and
processing), their applications, and research timeline. In addition, we discuss trends and opportunities
of IoT applications for smart agriculture and also indicate the open issues and challenges of IoT
application in smart agriculture. We hope that the findings of this study will constitute important
guidelines in research and promotion of IoT solutions aiming to improve the productivity and quality
of the agriculture sector as well as fa